Fix possible bootstrap issue
When 2 nodes were started for the first time whereas:
- one of them is in client-only mode, connected to the other one
- the registry node is temporarily down
then the normal node never tried to rebootstrap or connect directly to the
only node it knows (the registry node).
Such case required to restart the daemon when the registry is back.

Moreover, there was no reason to query the registry node immediately after
having open new tunnels to peers found in cache, when this number is less than
expected.