doc: use :Contact: instead of :Author: in the README
With :Author:, people may think I am the only author (which is not the case,
as shown by the Git history), whereas I just wanted to specify who is the lead
developer, i.e. the person to contact. And it's not a place to list here people
that are not involved anymore in the project.