Fix UDP for OpenVPN 2.4
OpenVPN 2.4 does not work anymore with proto 'udp', so we changed
to 'udp4' which is not compatible with 2.3, therefore, support for
OpenVPN 2.3 was dropped.
The link-mtu option changed since OpenVPN 2.3, causing mtu-size to
drop below the minimum with our chosen values.
This resulted in lack of link local IPv6 on udp interfaces.