Fix UDP for OpenVPN 2.4
OpenVPN 2.4 does not work anymore with proto 'udp', so we changed
to 'udp4' which is not compatible with 2.3, therefore, support for
OpenVPN 2.3 was dropped.
Strangely, there's no link-local IPv6 with 'udp4', as if OpenVPN
mixed up the payload with the underlying protocol.
Fortunately, we can rely on the kernel dual-stack.