Fix UDP for OpenVPN 2.4
OpenVPN 2.4 does not work anymore with proto 'udp', so we changed
to 'udp4' which is not compatible with 2.3, therefore, support for
OpenVPN 2.3 was dropped.
Also, we don't get a link local IPv6 address on udp interfaces if we use
udp/udp4 protocol on server side, so we rely on kernel automatic
translation and force server to pass --proto udp6.