demo: add testing of UDP
The verification of both UDP6 and UDP4 is added to the demo by
forcing machine 8 to create a monotunnel, either with machine 6 or 7.
These 2 machines, additionnaly to talking to machine8 with UDP6,
can talk to the registry and to each other via UDP4, they also keep
their ability to create tunnels via TCP with other machines.