fix disable-proto option
The default value for disable-proto is {'udp', 'udp6'}, and it's
not possible to get { <proto> } as a final value for disable-proto
when passing `disable-proto <proto>`, because this results in
disable_proto being equal to {'udp', 'udp6', <proto>}.
Similarly, passing both `disable-proto <proto>`
and `disable-proto none` results in the `none` overwriting anything
else, resulting in disable_proto being an empty list.
The goal here is to allow to only keep the protocol we want to disable
when passing `disable-proto <proto>`, the default 'udp' and 'udp6'
are in this case removed from disable_proto.