Implement HMAC for babel
HMAC is added to babel to make sure nodes from a given re6st network
don't talk to nodes from another re6st network. This is useful when
machines from separate re6st networks are on a LAN.
The key is the same for all nodes with the same registry: a random
part created by their registry and passed through network parameters,
combined with the prefix and prefix length of this re6st network.

This uses the hmac branch of nexedi/babeld, based on
the WIP branch of jech/babeld with Nexedi patches and another patch
adding an option to not check HMAC.